Move new UI into root components folder
