+++ web/src/components/ErrorBanner.vue
