+++ web/src/components/StreamData.vue
