+++ web/src/components/searchHistory.js
